+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Add `add_path` to proc config
 - Highlight changed unselected processes
 - Keymap help now uses actual keys (respecting config)
 - Clears the terminal before the first render.